pyproject.toml
[project]
name = "compvision-ml-movsport"
version = "0.1.0"
description = "Scripts didáticos de Visão Computacional e CNN para rastreamento de ponto (LaBioCoM / USP)"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
    { name = "Paulo Roberto Pereira Santiago", email = "" }
]
keywords = ["computer-vision", "deep-learning", "pytorch", "sports-science", "biomechanics"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]

dependencies = [
    "numpy>=1.20",
    "pillow>=9.0",
    "torch>=2.0",
    "matplotlib>=3.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]
